+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,9 @@
 --->
 
 ## Upcoming Release
+
+## Version 7.26.0
+* Onboarded `Microsoft.KubernetesConfiguration/privateLinkScopes` to Private Link Common Cmdlets
 * Added EdgeZone (Extended Location) parameter support for Azure Firewall
     - Added `-EdgeZone` parameter to `New-AzFirewall` cmdlet
     - When `-EdgeZone` is specified, availability zones are not supported for Azure Firewall resources